+++ packages/create-kumiko-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # create-kumiko-app
 
+## 0.285.2
+
+### Patch Changes
+
+- Updated dependencies [9c28242]
+  - @cosmicdrift/kumiko-framework@0.285.2
+  - @cosmicdrift/kumiko-dev-server@0.285.2
+  - @cosmicdrift/kumiko-server-runtime@0.285.2
+
 ## 0.285.1
 
 ### Patch Changes